Create new Grid to replace old monolithic Grid
